fix(hr): map tax rates as BigDecimal, and stop the test harness hiding schema drift
hr-service could not start against any real migrated database. tax_config's three rate
columns are NUMERIC(6,3) (010-create-hr-tables.xml) while TaxConfigEntity mapped them as
`double` (float8), and the deployed configuration runs ddl-auto: validate — so Hibernate
rejected the mapping at startup. Found while executing 13-03, which had to weaken its
HR gating assertions because the service would not come up.

All 18 integration tests passed regardless, because HrTestBase overrode ddl-auto to
"none". The harness switched off the single check that compares the JPA mappings to the
schema Liquibase had just built, so a green suite was reporting on a service that could
not boot. That override is now "validate", matching the deployed configuration. It is the
more valuable half of this commit: the type error is one bug, the blind spot would have
produced the next one.

BigDecimal is also the correct type independently. These are statutory tax and
contribution rates and binary floating point cannot represent values like 0.001 exactly.
PercentOfPaisa states the rate-to-paisa rounding rule once, HALF_UP, which is what
Math.round(x / 100.0) already did for the non-negative amounts payroll deals in — so no
existing payroll figure moves. Rounding is now explicit rather than a property of the
arithmetic, and a null rate throws instead of silently computing zero.

Impact analysis before the change: EobiCalculator.employeeContribution and
SlabTaxCalculator.computeSurcharge each have exactly one caller, PayrollRunService.calculate,
LOW risk, one affected process.

Verified: mvn -pl services/hr-service verify with ddl-auto=validate -> 30 tests, 0 errors,
0 failures (so tax_config was the only drift). hr-service then booted against the live dev
hr_db — "Started HrServiceApplication in 5.97 seconds" — and /actuator/health returns 200.

+++ services/hr-service/src/main/java/io/restaurantos/hr/entity/TaxConfigEntity.java
@@ -10,6 +10,7 @@ import jakarta.persistence.Table;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
@@ -47,14 +48,23 @@ public class TaxConfigEntity {
     @Column(name = "surcharge_threshold_paisa", nullable = false)
     private long surchargeThresholdPaisa;
 
+    // BigDecimal, not double. These three columns are NUMERIC(6,3) (010-create-hr-tables.xml), and
+    // Hibernate runs with ddl-auto: validate — a double maps to float8, so schema validation failed
+    // and hr-service could not start at all against a real migrated database. Every integration
+    // test passed regardless because HrTestBase overrides ddl-auto to "none", which switches off
+    // the one check that would have caught this.
+    //
+    // The type is also the correct one on its own merits: these are statutory tax and contribution
+    // rates, and binary floating point cannot represent values like 0.001 exactly. Keeping them
+    // decimal end-to-end means the rate that was configured is the rate that is applied.
     @Column(name = "surcharge_rate_pct", nullable = false)
-    private double surchargeRatePct;
+    private BigDecimal surchargeRatePct;
 
     @Column(name = "eobi_employer_rate_pct", nullable = false)
-    private double eobiEmployerRatePct;
+    private BigDecimal eobiEmployerRatePct;
 
     @Column(name = "eobi_employee_rate_pct", nullable = false)
-    private double eobiEmployeeRatePct;
+    private BigDecimal eobiEmployeeRatePct;
 
     @Column(name = "eobi_wage_base_paisa", nullable = false)
     private long eobiWageBasePaisa;
@@ -96,15 +106,15 @@ public class TaxConfigEntity {
         return surchargeThresholdPaisa;
     }
 
-    public double getSurchargeRatePct() {
+    public BigDecimal getSurchargeRatePct() {
         return surchargeRatePct;
     }
 
-    public double getEobiEmployerRatePct() {
+    public BigDecimal getEobiEmployerRatePct() {
         return eobiEmployerRatePct;
     }
 
-    public double getEobiEmployeeRatePct() {
+    public BigDecimal getEobiEmployeeRatePct() {
         return eobiEmployeeRatePct;
     }
